# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Task API",
        "description": "A simple REST API for managing tasks.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "summary": "Get API information",
                "responses": {
                    "200": {
                        "description": "API information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Task API"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "endpoints": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "/tasks"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "summary": "Say hello",
                "responses": {
                    "200": {
                        "description": "Hello message",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Hello World!"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Check API health",
                "responses": {
                    "200": {
                        "description": "API is healthy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "summary": "Get tasks",
                "description": "Returns all tasks, optionally filtered by completion status and title search.",
                "parameters": [
                    {
                        "name": "done",
                        "in": "query",
                        "required": false,
                        "description": "Filter tasks by completion status.",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "description": "Search for tasks by title. The search is case-insensitive and matches partial titles.",
                        "schema": {
                            "type": "string",
                            "example": "node"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "invalidDone": {
                                        "summary": "Invalid done value",
                                        "value": {
                                            "error": "done must be true or false"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new task",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            },
                            "example": {
                                "title": "Learn OpenAPI"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid task title",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Title cannot be empty"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "summary": "Get a task by ID",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 999 not found"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a task",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            },
                            "example": {
                                "title": "Learn OpenAPI properly",
                                "done": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid task data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 999 not found"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a task",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskId"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted successfully"
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 999 not found"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "TaskId": {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "The ID of the task",
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "schemas": {
            "Task": {
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "done"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique task ID",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "Task title",
                        "example": "Learn OpenAPI"
                    },
                    "done": {
                        "type": "boolean",
                        "description": "Whether the task is completed",
                        "example": false
                    }
                }
            },
            "CreateTaskRequest": {
                "type": "object",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Title of the new task",
                        "minLength": 1,
                        "example": "Learn OpenAPI"
                    }
                }
            },
            "UpdateTaskRequest": {
                "type": "object",
                "required": [
                    "title",
                    "done"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Updated task title",
                        "minLength": 1,
                        "example": "Learn OpenAPI properly"
                    },
                    "done": {
                        "type": "boolean",
                        "description": "Whether the task is completed",
                        "example": true
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Task 999 not found"
                    }
                }
            }
        }
    }
}